+++ src/services/errors/index.ts
@@ -1,0 +1,8 @@
+/**
+ * Error Handling System Exports
+ */
+
+export * from './ErrorClassification';
+export * from './ErrorLogger';
+export * from './ErrorDisplay';
+export * from './RetryMechanism';